fix DTO models and settings.gradle.kts

+++ music-api/src/main/java/dev/ginger/musicapi/models/AlbumDTO.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class AlbumDTO(
-    @SerialName("id") val id: Int,
+    @SerialName("id") val id: Long,
     @SerialName("title") val title: String,
     @SerialName("cover_small") val coverSmall: String,
     @SerialName("cover_medium") val coverMedium: String,
